backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
scikit-learn==1.3.2
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
xgboost==2.0.3
python-dotenv==1.0.1
scipy>=1.11.4
drf-yasg==1.21.7 